Extract namespace
